backend/server/requirements.txt
flask
flask_cors
PyJWT==1.7.1
python-dotenv
mysql-connector-python